_archive/legacy_core_backup/requirements-dev.txt

# 🧪 Phase 6: Visionary Auditor (Dev Only)
playwright==1.57.0
pytest-playwright==0.7.2
Pillow==10.2.0
numpy==1.26.4
